+++ lsql-core/src/main/java/com/w11k/lsql/converter/ConverterRegistry.java
@@ -52,6 +52,10 @@ public class ConverterRegistry {
     }
 
     public void addSqlToJavaConverter(Converter converter, boolean replaceExisting) {
+        if (converter.isWriteOnly()) {
+            return;
+        }
+
         if (!replaceExisting && this.sqlToJavaConverters.containsKey(converter.getSqlType())) {
             throw new IllegalStateException(
                     "A converter for the SQL type "