+++ lsql-core/src/main/java/com/w11k/lsql/converter/types/ByteConverter.java
@@ -1,4 +1,4 @@
-package com.w11k.lsql.converter.sqltypes;
+package com.w11k.lsql.converter.types;
 
 import com.w11k.lsql.LSql;
 import com.w11k.lsql.converter.Converter;
@@ -8,25 +8,19 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Types;
 
-public class StringConverter extends Converter {
+public class ByteConverter extends Converter {
 
-    public static int[] SQL_TYPES = new int[]{
-            Types.CHAR, Types.VARCHAR, Types.LONGNVARCHAR, Types.LONGVARCHAR, Types.NCHAR, Types.NVARCHAR
-    };
-
-
-    public StringConverter(int sqlType) {
-        super(String.class, sqlType);
+    public ByteConverter() {
+        super(com.w11k.lsql.Blob.class, Types.BINARY);
     }
 
     @Override
     protected void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
-        ps.setString(index, val.toString());
-
+        ps.setByte(index, (Byte) val);
     }
 
     @Override
     protected Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
-        return rs.getString(index);
+        return rs.getByte(index);
     }
 }