+++ lsql-core/src/main/java/com/w11k/lsql/converter/types/DoubleConverter.java
@@ -1,4 +1,4 @@
-package com.w11k.lsql.converter.sqltypes;
+package com.w11k.lsql.converter.types;
 
 import com.w11k.lsql.LSql;
 import com.w11k.lsql.converter.Converter;
@@ -8,20 +8,24 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Types;
 
-public class FloatConverter extends Converter {
+public class DoubleConverter extends Converter {
 
-    public FloatConverter() {
-        super(Float.class, Types.FLOAT);
+    public static int[] SQL_TYPES = new int[]{
+            Types.DOUBLE, Types.REAL, Types.DECIMAL, Types.NUMERIC
+    };
+
+    public DoubleConverter(int sqlType) {
+        super(Double.class, sqlType);
     }
 
     @Override
     protected void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
-        ps.setFloat(index, (Integer) val);
+        ps.setDouble(index, (Double) val);
     }
 
     @Override
     protected Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
-        return rs.getFloat(index);
+        return rs.getDouble(index);
 
     }
 }