+++ lsql-core/src/main/java/com/w11k/lsql/converter/types/IntConverter.java
@@ -1,4 +1,4 @@
-package com.w11k.lsql.converter.sqltypes;
+package com.w11k.lsql.converter.types;
 
 import com.w11k.lsql.LSql;
 import com.w11k.lsql.converter.Converter;
@@ -8,20 +8,24 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Types;
 
-public class LongConverter extends Converter {
+public class IntConverter extends Converter {
 
-    public LongConverter() {
-        super(Long.class, Types.BIGINT);
+    public static int[] SQL_TYPES = new int[]{
+            Types.TINYINT, Types.SMALLINT, Types.INTEGER
+    };
+
+    public IntConverter(int sqlType) {
+        super(Integer.class, sqlType);
     }
 
     @Override
     protected void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
-        ps.setLong(index, (Long) val);
+        ps.setInt(index, (Integer) val);
     }
 
     @Override
     protected Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
-        return rs.getLong(index);
+        return rs.getInt(index);
 
     }
 }