+++ lsql-core/src/main/java/com/w11k/lsql/converter/types/JodaDateTimeConverter.java
@@ -1,21 +1,21 @@
-package com.w11k.lsql.converter.sqltypes;
+package com.w11k.lsql.converter.types;
 
 import com.w11k.lsql.LSql;
 import com.w11k.lsql.converter.Converter;
-import org.joda.time.LocalDate;
+import org.joda.time.DateTime;
 
 import java.sql.*;
 
-public class JodaLocalDateConverter extends Converter {
+public class JodaDateTimeConverter extends Converter {
 
-    public JodaLocalDateConverter() {
-        super(LocalDate.class, Types.DATE);
+    public JodaDateTimeConverter() {
+        super(DateTime.class, Types.TIMESTAMP);
     }
 
     @Override
     protected void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
-        LocalDate dt = (LocalDate) val;
-        Timestamp ts = new Timestamp(dt.toDate().getTime());
+        DateTime dt = (DateTime) val;
+        Timestamp ts = new Timestamp(dt.getMillis());
         ps.setTimestamp(index, ts);
     }
 
@@ -23,7 +23,7 @@ public class JodaLocalDateConverter extends Converter {
     protected Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
         Timestamp timestamp = rs.getTimestamp(index);
         if (timestamp != null) {
-            return new LocalDate(timestamp.getTime());
+            return new DateTime(timestamp.getTime());
         } else {
             return null;
         }